+++ packages/react-charting/src/components/HorizontalBarChartWithAxis/HorizontalBarChartWithAxis.types.ts
@@ -8,7 +8,7 @@ import {
 } from '../../index';
 
 /**
- * IHorizontalBarChartWithAxisProps for HorizontalBarchartWithAxis component.
+ * Horizontal Bar Chart with Axis properties
  * {@docCategory HorizontalBarChartWithAxis}
  */
 export interface IHorizontalBarChartWithAxisProps extends ICartesianChartProps {
@@ -72,6 +72,10 @@ export interface IHorizontalBarChartWithAxisProps extends ICartesianChartProps {
   showYAxisLables?: boolean;
 }
 
+/**
+ * Horizontal Bar Chart with Axis style properties
+ * {@docCategory HorizontalBarChartWithAxis}
+ */
 export interface IHorizontalBarChartWithAxisStyleProps extends ICartesianChartStyleProps {
   /**
    * color of the datapoint legend
@@ -79,6 +83,10 @@ export interface IHorizontalBarChartWithAxisStyleProps extends ICartesianChartSt
   legendColor?: string;
 }
 
+/**
+ * Horizontal Bar Chart with Axis styles
+ * {@docCategory HorizontalBarChartWithAxis}
+ */
 export interface IHorizontalBarChartWithAxisStyles extends ICartesianChartStyles {
   /**
    * Style for the chart label.