+++ packages/react-charting/src/components/VerticalBarChart/VerticalBarChart.tsx
@@ -6,7 +6,7 @@ import { getStyles } from './VerticalBarChart.styles';
 
 // Create a VerticalBarChart variant which uses these default styles and this styled subcomponent.
 /**
- * VerticalBarchart component.
+ * VerticalBarchart component
  * {@docCategory VerticalBarChart}
  */
 export const VerticalBarChart: React.FunctionComponent<IVerticalBarChartProps> = styled<