+++ packages/react-components/react-timepicker-compat-preview/src/components/TimePicker/TimePicker.types.ts
@@ -1,3 +1,4 @@
+import * as React from 'react';
 import type { ComboboxSlots, ComboboxState, ComboboxProps, SelectionEvents } from '@fluentui/react-combobox';
 
 export type Hour =
@@ -47,11 +48,44 @@ export type TimePickerOption = {
   text: string;
 };
 
+/**
+ * Error types returned by the `onValidationResult` callback.
+ */
+export type TimePickerErrorType = 'invalid-input' | 'out-of-bounds';
+
+export type TimeStringValidationResult = {
+  date: Date | null;
+  error?: TimePickerErrorType;
+};
+
 export type TimePickerSlots = ComboboxSlots;
 
-export type TimeSelectionEvents = SelectionEvents;
+export type TimeSelectionEvents = SelectionEvents | React.FocusEvent<HTMLElement>;
 export type TimeSelectionData = {
-  selectedTime: Date | undefined;
+  /**
+   * The Date object associated with the selected option. For freeform TimePicker it can also be the Date object parsed from the user input.
+   */
+  selectedTime: Date | null;
+  /**
+   * The display text for the selected option. For freeform TimePicker it can also be the value in user input.
+   */
+  selectedTimeText: string | undefined;
+  /**
+   * The error type for the selected option.
+   */
+  error: TimePickerErrorType | undefined;
+};
+
+export type TimeFormatOptions = {
+  /**
+   * If true, use 12-hour time format. Otherwise, use 24-hour format.
+   */
+  hour12?: boolean;
+
+  /**
+   * If true, show seconds in the dropdown options and consider seconds for default validation purposes.
+   */
+  showSeconds?: boolean;
 };
 
 /**
@@ -66,54 +100,61 @@ export type TimePickerProps = Omit<
   | 'multiselect'
   | 'onOptionSelect'
   | 'selectedOptions'
-> & {
-  /**
-   * If true, use 12-hour time format. Otherwise, use 24-hour format.
-   */
-  hour12?: boolean;
+> &
+  TimeFormatOptions & {
+    /**
+     * Start hour (inclusive) for the time range, 0-24.
+     */
+    startHour?: Hour;
 
-  /**
-   * Start hour (inclusive) for the time range, 0-24.
-   */
-  startHour?: Hour;
+    /**
+     * End hour (exclusive) for the time range, 0-24.
+     */
+    endHour?: Hour;
 
-  /**
-   * End hour (exclusive) for the time range, 0-24.
-   */
-  endHour?: Hour;
+    /**
+     * Time increment, in minutes, of the options in the dropdown.
+     */
+    increment?: number;
 
-  /**
-   * Time increment, in minutes, of the options in the dropdown.
-   */
-  increment?: number;
+    /**
+     * The date in which all dropdown options are based off of.
+     */
+    dateAnchor?: Date;
 
-  /**
-   * The date in which all dropdown options are based off of.
-   */
-  dateAnchor?: Date;
+    /**
+     * Currently selected time in the TimePicker.
+     */
+    selectedTime?: Date | null;
 
-  /**
-   * If true, show seconds in the dropdown options and consider seconds for default validation purposes.
-   */
-  showSeconds?: boolean;
+    /**
+     * Default selected time in the TimePicker, for uncontrolled scenarios.
+     */
+    defaultSelectedTime?: Date;
 
-  /**
-   * Currently selected time in the TimePicker.
-   */
-  selectedTime?: Date;
+    /**
+     * Callback for when a time selection is made.
+     */
+    onTimeSelect?: (event: TimeSelectionEvents, data: TimeSelectionData) => void;
 
-  /**
-   * Default selected time in the TimePicker, for uncontrolled scenarios.
-   */
-  defaultSelectedTime?: Date;
+    /**
+     * Custom the date strings displayed (in dropdown options and input).
+     */
+    formatDateToTimeString?: (date: Date) => string;
 
-  /**
-   * Callback for when a time selection is made.
-   */
-  onTimeSelect?: (event: TimeSelectionEvents, data: TimeSelectionData) => void;
-};
+    /**
+     * Custom validation for the input time string from user in freeform TimePicker.
+     */
+    validateFreeFormTime?: (time: string | undefined) => TimeStringValidationResult;
+  };
 
 /**
  * State used in rendering TimePicker
  */
-export type TimePickerState = ComboboxState;
+export type TimePickerState = ComboboxState &
+  Required<Pick<TimePickerProps, 'freeform' | 'validateFreeFormTime'>> & {
+    /**
+     * Submitted text from the input field. It is used to determine if the input value has changed when user submit a new value on Enter or blur from input.
+     */
+    submittedText: string | undefined;
+  };