+++ packages/react-components/react-timepicker-compat-preview/src/components/TimePicker/useTimePicker.tsx
@@ -1,14 +1,16 @@
 import * as React from 'react';
-import { useControllableState } from '@fluentui/react-utilities';
+import { elementContains, mergeCallbacks, useControllableState, useMergedRefs } from '@fluentui/react-utilities';
+import { Enter } from '@fluentui/keyboard-keys';
 import type { Hour, TimePickerOption, TimePickerProps, TimePickerState, TimeSelectionData } from './TimePicker.types';
 import { ComboboxProps, useCombobox_unstable, Option } from '@fluentui/react-combobox';
 import {
   dateToKey,
   keyToDate,
-  formatTimeString,
+  formatDateToTimeString as defaultFormatDateToTimeString,
   getDateStartAnchor,
   getDateEndAnchor,
   getTimesBetween,
+  getDateFromTimeString,
 } from './timeMath';
 
 // TODO before stable, replace useCallback to useEventCallback if needed
@@ -27,14 +29,17 @@ export const useTimePicker_unstable = (props: TimePickerProps, ref: React.Ref<HT
     dateAnchor: dateAnchorInProps,
     defaultSelectedTime: defaultSelectedTimeInProps,
     endHour = 24,
+    formatDateToTimeString,
     hour12 = false,
     increment = 30,
     onTimeSelect,
     selectedTime: selectedTimeInProps,
     showSeconds = false,
     startHour = 0,
+    validateFreeFormTime: validateFreeFormTimeInProps,
     ...rest
   } = props;
+  const { freeform = false } = rest;
 
   const { dateStartAnchor, dateEndAnchor } = useStableDateAnchor(
     dateAnchorInProps ?? selectedTimeInProps ?? defaultSelectedTimeInProps,
@@ -42,21 +47,39 @@ export const useTimePicker_unstable = (props: TimePickerProps, ref: React.Ref<HT
     endHour,
   );
 
+  const dateToText = React.useCallback(
+    (dateTime: Date) =>
+      formatDateToTimeString
+        ? formatDateToTimeString(dateTime)
+        : defaultFormatDateToTimeString(dateTime, { showSeconds, hour12 }),
+    [hour12, formatDateToTimeString, showSeconds],
+  );
   const options: TimePickerOption[] = React.useMemo(
     () =>
       getTimesBetween(dateStartAnchor, dateEndAnchor, increment).map(time => ({
         date: time,
         key: dateToKey(time),
-        text: formatTimeString(time, showSeconds, hour12),
+        text: dateToText(time),
       })),
-    [dateStartAnchor, dateEndAnchor, increment, showSeconds, hour12],
+    [dateStartAnchor, dateEndAnchor, increment, dateToText],
   );
 
-  const [selectedTime, setSelectedTime] = useControllableState<Date | undefined>({
-    state: props.selectedTime,
-    defaultState: props.defaultSelectedTime,
-    initialState: undefined,
+  const [selectedTime, setSelectedTime] = useControllableState<Date | null>({
+    state: selectedTimeInProps,
+    defaultState: defaultSelectedTimeInProps,
+    initialState: null,
   });
+
+  const [submittedText, setSubmittedText] = React.useState<string | undefined>(undefined);
+
+  const selectTime: TimePickerProps['onTimeSelect'] = React.useCallback(
+    (e, data) => {
+      setSelectedTime(data.selectedTime);
+      setSubmittedText(data.selectedTimeText);
+      onTimeSelect?.(e, data);
+    },
+    [onTimeSelect, setSelectedTime],
+  );
 
   const selectedOptions = React.useMemo(() => {
     const selectedOption = options.find(date => date.key === dateToKey(selectedTime));
@@ -65,14 +88,21 @@ export const useTimePicker_unstable = (props: TimePickerProps, ref: React.Ref<HT
 
   const handleOptionSelect: ComboboxProps['onOptionSelect'] = React.useCallback(
     (e, data) => {
-      const timeSelectionData: TimeSelectionData = { selectedTime: keyToDate(data.optionValue) };
-      onTimeSelect?.(e, timeSelectionData);
-      setSelectedTime(timeSelectionData.selectedTime);
+      if (freeform && data.optionValue === undefined) {
+        // Combobox clears selection when input value not matching any option; but we allow this case in freeform TimePicker.
+        return;
+      }
+      const timeSelectionData: TimeSelectionData = {
+        selectedTime: keyToDate(data.optionValue),
+        selectedTimeText: data.optionText,
+        error: undefined,
+      };
+      selectTime(e, timeSelectionData);
     },
-    [setSelectedTime, onTimeSelect],
+    [freeform, selectTime],
   );
 
-  const state = useCombobox_unstable(
+  const baseState = useCombobox_unstable(
     {
       ...rest,
       selectedOptions,
@@ -86,6 +116,20 @@ export const useTimePicker_unstable = (props: TimePickerProps, ref: React.Ref<HT
     ref,
   );
 
+  const defaultValidateTime = React.useCallback(
+    (time: string | undefined) => getDateFromTimeString(time, dateStartAnchor, dateEndAnchor, { hour12, showSeconds }),
+    [dateEndAnchor, dateStartAnchor, hour12, showSeconds],
+  );
+
+  const state: TimePickerState = {
+    ...baseState,
+    freeform,
+    validateFreeFormTime: validateFreeFormTimeInProps ?? defaultValidateTime,
+    submittedText,
+  };
+
+  useSelectTimeFromValue(state, selectTime);
+
   return state;
 };
 
@@ -97,7 +141,7 @@ const useStableDateAnchor = (providedDate: Date | undefined, startHour: Hour, en
   const [fallbackDateAnchor] = React.useState(() => new Date());
 
   // Convert the Date object to a stable key representation. This ensures that the memoization remains stable when a new Date object representing the same date is passed in.
-  const dateAnchorKey = dateToKey(providedDate);
+  const dateAnchorKey = dateToKey(providedDate ?? null);
   const dateAnchor = React.useMemo(
     () => keyToDate(dateAnchorKey) ?? fallbackDateAnchor,
     [dateAnchorKey, fallbackDateAnchor],
@@ -110,4 +154,72 @@ const useStableDateAnchor = (providedDate: Date | undefined, startHour: Hour, en
   );
 
   return { dateStartAnchor, dateEndAnchor };
+};
+
+/**
+ * Mimics the behavior of the browser's change event for a freeform TimePicker.
+ * The provided callback is called when input changed and:
+ * - Enter/Tab key is pressed on the input.
+ * - TimePicker loses focus, signifying a possible change.
+ */
+const useSelectTimeFromValue = (state: TimePickerState, callback: TimePickerProps['onTimeSelect']) => {
+  const { activeOption, freeform, validateFreeFormTime, options, submittedText, setActiveOption, value } = state;
+
+  // Base Combobox has activeOption default to first option in dropdown even if it doesn't match input value, and Enter key will select it.
+  // This effect ensures that the activeOption is cleared when the input doesn't match any option.
+  // This behavior is specific to a freeform TimePicker where the input value is treated as a valid time even if it's not in the dropdown.
+  const isValueOptionPrefix = value ? options.some(({ text }) => text.indexOf(value) === 0) : false;
+  React.useEffect(() => {
+    if (freeform && value && !isValueOptionPrefix) {
+      setActiveOption(undefined);
+    }
+  }, [freeform, isValueOptionPrefix, setActiveOption, value]);
+
+  const selectTimeFromValue = React.useCallback(
+    (e: React.KeyboardEvent<HTMLInputElement> | React.FocusEvent<HTMLInputElement>) => {
+      if (!freeform) {
+        return;
+      }
+
+      const { date: selectedTime, error } = validateFreeFormTime(value);
+
+      // Only triggers callback when the text in input has changed.
+      if (submittedText !== value) {
+        callback?.(e, { selectedTime, selectedTimeText: value, error });
+      }
+    },
+    [callback, freeform, submittedText, validateFreeFormTime, value],
+  );
+
+  const handleKeyDown: ComboboxProps['onKeyDown'] = React.useCallback(
+    e => {
+      if (!activeOption && e.key === Enter) {
+        selectTimeFromValue(e);
+      }
+    },
+    [activeOption, selectTimeFromValue],
+  );
+  state.root.onKeyDown = mergeCallbacks(handleKeyDown, state.root.onKeyDown);
+
+  const rootRef = React.useRef<HTMLDivElement>(null);
+  state.root.ref = useMergedRefs(state.root.ref, rootRef);
+
+  if (state.listbox) {
+    state.listbox.tabIndex = -1; // allows it to be the relatedTarget of a blur event.
+  }
+
+  if (state.expandIcon) {
+    state.expandIcon.tabIndex = -1; // allows it to be the relatedTarget of a blur event.
+  }
+
+  const handleInputBlur = React.useCallback(
+    (e: React.FocusEvent<HTMLInputElement>) => {
+      const isOutside = e.relatedTarget ? !elementContains(rootRef.current, e.relatedTarget) : true;
+      if (isOutside) {
+        selectTimeFromValue(e);
+      }
+    },
+    [selectTimeFromValue],
+  );
+  state.input.onBlur = mergeCallbacks(handleInputBlur, state.input.onBlur);
 };